pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer - SEO content analysis, third project of Hexlet's Python Course"
authors = ["phelix alienflakes <alienflakes38@gmail.com>"]
packages = [
    { include = "page_analyzer" }
]


[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
